webhook-repo/requirements.txt
Flask==2.3.2
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0
black==23.7.0
flake8==6.0.0
isort==5.12.0
safety==2.3.5
